Rename traits to follow Laravel conventions
