+++ src/Model/Term.php
@@ -3,8 +3,8 @@
 namespace Corcel\Model;
 
 use Corcel\Model;
-use Corcel\Traits\HasAcfFields;
-use Corcel\Traits\HasMetaFields;
+use Corcel\Corcerns\AdvancedCustomFields;
+use Corcel\Corcerns\MetaFields;
 
 /**
  * Class Term.
@@ -14,8 +14,8 @@ use Corcel\Traits\HasMetaFields;
  */
 class Term extends Model
 {
-    use HasMetaFields;
-    use HasAcfFields;
+    use MetaFields;
+    use AdvancedCustomFields;
 
     /**
      * @var string
